* Added the possibility to see and edit the sniffer rule of a type   (realized as an option enabled in the menu). * Minor bug fixes and changes.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@16791 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/preferences/filetypes/StringView.cpp
@@ -173,7 +173,7 @@ StringView::GetPreferredSize(float* _width, float* _height)
 	if (_height) {
 		font_height fontHeight;
 		GetFontHeight(&fontHeight);
-		*_height = ceil(fontHeight.ascent + fontHeight.descent + fontHeight.leading) + 2.0f;
+		*_height = ceilf(fontHeight.ascent + fontHeight.descent + fontHeight.leading) + 2.0f;
 	}
 }
